Fail if no version is given to deploy.

+++ zero.py
@@ -75,6 +75,12 @@ class ZeroApp:
     RUN => 'find /opt/rlci -mindepth 1 -maxdepth 1 -exec rm -rf {} \\\\;'
     RUN => 'git clone git@github.com:rickardlindberg/rlci.git /opt/rlci/'
     RUN => 'git -C /opt/rlci checkout <git-hash>'
+
+    I fail if no version is given:
+
+    >>> ZeroApp.run_in_test_mode(args=['deploy'])
+    STDOUT => 'No version given to deploy.'
+    EXIT => 1
     """
 
     def __init__(self, args, terminal, tests, shell):
@@ -101,6 +107,9 @@ class ZeroApp:
             self.shell.run("git branch -d BRANCH")
             self.shell.run("git push origin BRANCH --delete")
         elif self.args.get()[:1] == ["deploy"]:
+            if len(self.args.get()) < 2:
+                self.terminal.print_line("No version given to deploy.")
+                sys.exit(1)
             version = self.args.get()[1]
             self.shell.run("find /opt/rlci -mindepth 1 -maxdepth 1 -exec rm -rf {} \;")
             self.shell.run("git clone git@github.com:rickardlindberg/rlci.git /opt/rlci/")
